+++ src/app/api/aging-index-vs-race/route.ts
@@ -3,7 +3,7 @@ import { query } from "@/lib/db";
 
 export async function GET() {
   const data = await query(
-    `SELECT cor_raca, ROUND(AVG(indice_envelhecimento), 2) AS indice
+    `SELECT cor_raca, ROUND(AVG(indice_envelhecimento)::numeric, 2) AS indice
 FROM indice_envelhecimento_raca
 WHERE ano = 2022
 GROUP BY cor_raca;`
